pyproject.toml
[tool.poetry]
name = "movie_tools"
version = "0.1.0"
description = ""
authors = ["shin-hama <htysivista@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
Pillow = "^9.2.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
mypy = "^0.961"
flake8 = "^4.0.1"
black = "^22.6.0"
types-requests = "^2.28.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
